+++ funsor/autodiff.py
